string: Fix strncasecmp signature

+++ include/strings.h
@@ -25,7 +25,7 @@ extern int strcasecmp(const char *str1, const char *str2);
 
 
 /* Compares at most the first n bytes of str1 and str2 case-insensitively. */
-extern int strncasecmp(const char *s1, const char *s2, int n);
+extern int strncasecmp(const char *s1, const char *s2, size_t n);
 
 
 #endif
